+++ subsys/mgmt/mcumgr/smp_internal.h
@@ -8,13 +8,31 @@
 #ifndef MGMT_MCUMGR_SMP_INTERNAL_H_
 #define MGMT_MCUMGR_SMP_INTERNAL_H_
 
+#include <stdint.h>
+#include <zephyr/net/buf.h>
+#include <zephyr/mgmt/mcumgr/smp.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
+struct mgmt_hdr {
+#ifdef CONFIG_LITTLE_ENDIAN
+	uint8_t  nh_op:3;		/* MGMT_OP_[...] */
+	uint8_t  _res1:5;
+#else
+	uint8_t  _res1:5;
+	uint8_t  nh_op:3;		/* MGMT_OP_[...] */
+#endif
+	uint8_t  nh_flags;		/* Reserved for future flags */
+	uint16_t nh_len;		/* Length of the payload */
+	uint16_t nh_group;		/* MGMT_GROUP_ID_[...] */
+	uint8_t  nh_seq;		/* Sequence number */
+	uint8_t  nh_id;			/* Message ID within group */
+};
+
 struct smp_transport;
 struct zephyr_smp_transport;
-struct net_buf;
 
 /**
  * @brief Enqueues an incoming SMP request packet for processing.
